+++ CHANGELOG.md
@@ -17,7 +17,19 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 <!-- ### Fixed -->
 <!-- ### Removed -->
 
-<!-- ## Unreleased -->
+## Unreleased
+
+### Added
+
+- TypeScript type errors are now displayed inline. Previously, only syntax
+  errors were displayed. Type errors may take slightly longer to appear than
+  syntax errors, because they require fetching `.d.ts` files and running a more
+  expensive compiler pass.
+
+### Changed
+
+- [**BREAKING**] Removed `--playground-error-border` property, and updated
+  inline error style from a dashed border to a red squiggle.
 
 ## [0.7.0] - 2021-03-30